+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Next Release
 
-- [#944](https://github.com/IAMconsortium/pyam/pull/944) Refactor to a format_n() function for nice log messages
+- [#946](https://github.com/IAMconsortium/pyam/pull/946) Handle plot-styling by meta indicators with `np.nan`
+- [#944](https://github.com/IAMconsortium/pyam/pull/944) Refactor to a `format_n()` function for nice log messages
 - [#943](https://github.com/IAMconsortium/pyam/pull/943) Improved handling for division by zero
 - [#941](https://github.com/IAMconsortium/pyam/pull/941) Add a `compute.share()` method
 - [#938](https://github.com/IAMconsortium/pyam/pull/938) Add a function to initialize an IamDataFrame from an *